requirements.txt
fastapi
uvicorn[standard]
argon2-cffi
cryptography
pyjwt
sqlalchemy
python-dotenv
requests
beautifulsoup4
psycopg2-binary
redis
icecream
pydantic
pydantic[email]
python-multipart
setuptools
pytz
pandas
openpyxl
reportlab
pyrebase4
firebase-admin
pillow
apscheduler
orjson
brotli
brotli-asgi
asyncpg
gunicorn
jinja2
pyrebase4
itsdangerous
uvloop
